@@ -1,7 +1,5 @@
-module.exports = {
-  root: true,
-  extends: ['@liuchengjin/eslint-config/strict'],
-  rules: {
-    '@typescript-eslint/no-explicit-any': 'off',
-  },
-}
+import { defineFlatConfig } from 'eslint-define-config';
+import baseConfig from './src/base'
+export default defineFlatConfig( [
+  ...baseConfig
+])
